5/lesson_5/requirements.txt
python-decouple==3.8
termcolor==2.2.0
